requirements.txt
streamlit
mediapipe
opencv-python-headless
streamlit-webrtc
av
pillow
